+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # Changelog
 
+---
+
 ## What's new?
 
-This field is empty.
+The Int128 struct that works just like a regular Int32 or Int64.
 
-
+---
 
 ## Notes
+
+## Version 0.1.0 - 14/6/2026
+
+Add
+
+- Int128 struct. Arithmetical, equality and bitwise operators for the struct. ToByteArray conversion function.
 
 ### Version 0.0.0 - 13/6/2026